🚚 Rename: update package name and pom
