+++ care/src/main/java/com/animal/api/animal/service/UserAnimalServiceImple.java
@@ -8,7 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 
-import com.animal.api.animal.dao.UserAnimalMapper;
+import com.animal.api.animal.mapper.UserAnimalMapper;
 import com.animal.api.animal.model.response.AllAnimalListResponseDTO;
 
 @Service
