+++ care/src/main/resources/sql-mapper/animal/UserAnimalMapper.xml
@@ -3,7 +3,7 @@
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
   
-<mapper namespace="com.animal.api.animal.dao.UserAnimalMapper">
+<mapper namespace="com.animal.api.animal.mapper.UserAnimalMapper">
 <select id="getAllAnimals" parameterType="Map" resultType="com.animal.api.animal.model.response.AllAnimalListResponseDTO">
 SELECT
 	A.IDX,
