+++ care/src/main/resources/sql-mapper/donation/UserDonationsMapper.xml
@@ -2,8 +2,8 @@
 <!DOCTYPE mapper
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.animal.api.donation.dao.UserDonationsMapper">
-
+  
+<mapper namespace="com.animal.api.donation.mapper.UserDonationsMapper">
 <select id="getAllDonations" parameterType="Map" resultType="com.animal.api.donation.model.response.AllDonationListResponseDTO">
 SELECT
 	D.IDX,
